fix: Separation of logic between components
