+++ rs-react-app/eslint.config.js
@@ -35,7 +35,7 @@ export default tseslint.config(
       'react-compiler/react-compiler': 'error',
       ...react.configs.recommended.rules,
       ...react.configs['jsx-runtime'].rules,
-      "react/prop-types": "off"
+      'react/prop-types': 'off',
     },
     settings: {
       react: {
