+++ rs-react-app/src/components/inputElement/inputElem.module.css
@@ -1,7 +1,7 @@
 .search-header {
   background-color: #f8f9fa;
   padding: 2rem 0;
-  box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
   position: sticky;
   top: 0;
   z-index: 1;
