--- rs-react-app/src/components/resultsArea/resultsArea.tsx
@@ -1,4 +1,0 @@
-import { Component } from 'react';
-class ResultArea extends Component {}
-
-export default ResultArea;
